requirements.txt
flask
requests
uuid
flask-sqlalchemy
python-dotenv
psycopg2-binary
